% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemR.R
\name{align_telem}
\alias{align_telem}
\title{Take two tidy telem files and align(combine) them, without collpasing.
Note that new dates for tidy_telem_2 will be inaccurate.
If animals in the two experiments have the same number and meta data,
their values will be averaged by default}
\usage{
align_telem(
  tidy_telem_1,
  tidy_telem_2,
  align_date_1 = NULL,
  align_date_2 = NULL
)
}
\arguments{
\item{tidy_telem_1}{a tidy telemetry tibble, as produced by read_starr or read_oddi}

\item{tidy_telem_2}{a tidy telemetry tibble, as produced by read_starr or read_oddi}

\item{align_date_1}{date in tidy_telem_1 to align with date in tidy_telem_2}

\item{align_date_2}{date in tidy_telem_2 to align with date in tidy_telem_1}
}
\value{
a collpased tidy telemetry tibble
}
\description{
Take two tidy telem files and align(combine) them, without collpasing.
Note that new dates for tidy_telem_2 will be inaccurate.
If animals in the two experiments have the same number and meta data,
their values will be averaged by default
}
