% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemR.R
\name{combine_telem}
\alias{combine_telem}
\title{Take two tidy telem files and combine them,}
\usage{
combine_telem(tidy_telem_1, tidy_telem_2)
}
\arguments{
\item{tidy_telem_1}{a tidy telemetry tibble, as produced by read_starr or read_oddi}

\item{tidy_telem_2}{a tidy telemetry tibble, as produced by read_starr or read_oddi}
}
\value{
a collpased tidy telemetry tibble
}
\description{
Take two tidy telem files and combine them,
}
