% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemR.R
\name{export_telem}
\alias{export_telem}
\title{Read tidy telemetry tibbles and export them as a multi page excel file.}
\usage{
export_telem(tidy_telem, filename, return_list = F)
}
\arguments{
\item{tidy_telem}{a tidy telemetry tibble, as produced by read_starr or read_oddi}

\item{filename}{filename and path for saving excel file}

\item{return_list}{logical: if true, returns multi page object as a list}
}
\value{
list or null
}
\description{
Read tidy telemetry tibbles and export them as a multi page excel file.
}
