% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemR.R
\name{first_time}
\alias{first_time}
\title{return the earliest time in a time vector}
\usage{
first_time(time_vector)
}
\arguments{
\item{time_vector}{a vector of times in class posixct or anything that works with range()}
}
\value{
the earliest time in the vector
}
\description{
return the earliest time in a time vector
}
