% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemR.R
\name{graph_telem}
\alias{graph_telem}
\title{Graph tidy telemetry data.}
\usage{
graph_telem(
  tidy_telem,
  telem_var = "DegC",
  one_day_avg = F,
  group_by = "Mouse",
  tx_time = NULL,
  se_ribbon = T
)
}
\arguments{
\item{tidy_telem}{a tidy telemetry tibble, as produced by read_starr or read_oddi}

\item{telem_var}{telemetry measurement to plot.}

\item{one_day_avg}{collapse data into a 1 day average?}

\item{group_by}{variable to group by. Either "Group" or "Mouse"}

\item{tx_time}{vector of times (in posixct format) where to draw vertical dashed lines}

\item{se_ribbon}{logical: display standard error ribbon?}
}
\value{
plot
}
\description{
Graph tidy telemetry data.
}
