% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemR.R
\name{last_time}
\alias{last_time}
\title{return the latest time in a time vector}
\usage{
last_time(time_vector)
}
\arguments{
\item{time_vector}{a vector of times in class posixct or anything that works with range()}
}
\value{
the latest time in the vector
}
\description{
return the latest time in a time vector
}
