% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemR.R
\name{light_dark}
\alias{light_dark}
\title{add a logical column to tidy telem tibble indicating if lights are on (true)}
\usage{
light_dark(tidy_telem, lights_on_DST = 6)
}
\arguments{
\item{tidy_telem}{a tidy telemetry tibble, as produced by read_starr or read_oddi}

\item{lights_on_DST}{at what time do the lights in the room come on, during DST. This algorithm assumes that the room light schedule does not adjust for DST}
}
\value{
a tidy telemetry tibble with logical column "DST" added. Time is still clock time.
}
\description{
add a logical column to tidy telem tibble indicating if lights are on (true)
}
