% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemR.R
\name{lme_telem}
\alias{lme_telem}
\title{linear mixed effects model of telemetry data.}
\usage{
lme_telem(tidy_telem, telem_var = "DegC", collapse_first = T)
}
\arguments{
\item{tidy_telem}{a tidy telemetry tibble, as produced by read_starr or read_oddi}

\item{telem_var}{telemetry measurement to analyze}

\item{collapse_first}{logical. collapse to a 1 day average before analysis? recommended.}
}
\value{
list containing anova and models built
}
\description{
linear mixed effects model of telemetry data.
}
