% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemR.R
\name{MCsumm}
\alias{MCsumm}
\title{return asterisks from p values for easy reading}
\usage{
MCsumm(pval)
}
\arguments{
\item{pval}{a p value}
}
\value{
an asterisk code indicating significance
}
\description{
return asterisks from p values for easy reading
}
