% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemR.R
\name{trim_bad_probe}
\alias{trim_bad_probe}
\title{automatically detect and drop bad probes once they stop working. very useful for probes attached to mouse tails as they can fall off.}
\usage{
trim_bad_probe(tidy_telem, DegC_threshold = 23, over_how_long = "4 hours")
}
\arguments{
\item{tidy_telem}{a tidy telemetry tibble, as produced by read_starr or read_oddi}

\item{DegC_threshold}{what temperature in C is considered a spuriously low reading?}

\item{over_how_long}{temp values should be below the threshold over how big of a time window before probe is considered bad?}
}
\value{
a tidy telemetry tibble with bad probes removed starting a the point that they meet "bad" criteria
}
\description{
automatically detect and drop bad probes once they stop working. very useful for probes attached to mouse tails as they can fall off.
}
