% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemR.R
\name{trim_telem}
\alias{trim_telem}
\title{Trim tidy telemetry data.}
\usage{
trim_telem(
  tidy_telem,
  start_time = NULL,
  end_time = NULL,
  keep_groups = NULL,
  keep_mice = NULL,
  cut_mice = NULL
)
}
\arguments{
\item{tidy_telem}{a tidy telemetry tibble, as produced by read_starr or read_oddi}

\item{start_time}{start time for new output file}

\item{end_time}{end time for new output file}

\item{keep_groups}{optionally keep selected groups as defined in meta data. Default (NULL) is to keep all.}

\item{keep_mice}{optionally keep defined mice. Default (NULL) is to keep all.}
}
\value{
a tidy telemetry tibble
}
\description{
Trim tidy telemetry data.
}
