% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemR.R
\name{x_factor_telem}
\alias{x_factor_telem}
\title{Combine factors in a tidy telem giving a new column}
\usage{
x_factor_telem(tidy_telem, telem_var_1, telem_var_2)
}
\arguments{
\item{tidy_telem}{a tidy telemetry tibble, as produced by read_starr or read_oddi}

\item{telem_var_1}{first telem var to combine}

\item{telem_var_2}{second telem var to combine}
}
\value{
a  tidy telemetry tibble with a new column "x_factor"
}
\description{
Combine factors in a tidy telem giving a new column
}
