% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemR.R
\name{x_over_telem}
\alias{x_over_telem}
\title{Combine factors in a tidy telem giving a new column}
\usage{
x_over_telem(
  tidy_telem,
  day_1_start,
  day_2_start,
  xover_var = "Treated",
  xover_pattern_1 = "A",
  xover_pattern_2 = "B"
)
}
\arguments{
\item{tidy_telem}{a tidy telemetry tibble, as produced by read_starr or read_oddi. MUST HAVE grouping meta data in Xover}
}
\value{
a one day view of uncrossed telem data
}
\description{
Combine factors in a tidy telem giving a new column
}
