% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemR.R
\name{zeitgeber_time}
\alias{zeitgeber_time}
\title{convert time to zeitgeber (lightgiver) time: 0 = lights on, 12 = lights off.}
\usage{
zeitgeber_time(tidy_telem, undo = F, lights_on_DST = 7)
}
\arguments{
\item{tidy_telem}{a tidy telemetry tibble, as produced by read_starr or read_oddi}

\item{undo}{logical. if true, converts ZT to clock time.}

\item{lights_on_DST}{at what time do the lights in the room come on, during DST. This algorithm assumes that the room light schedule does not adjust for DST}
}
\value{
a tidy telemetry tibble converted to zeitgeber time.
}
\description{
convert time to zeitgeber (lightgiver) time: 0 = lights on, 12 = lights off.
}
