% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemR.R
\name{multcomp_telem}
\alias{multcomp_telem}
\title{multiple comparisons of telemetry data.}
\usage{
multcomp_telem(tidy_telem, formula, plot_or_table = "plot", collapse_first = T)
}
\arguments{
\item{tidy_telem}{a tidy telemetry tibble, as produced by read_starr or read_oddi}

\item{collapse_first}{optionally keep defined mice. Default (NULL) is to keep all}
}
\value{
list containing anova and models built
}
\description{
multiple comparisons of telemetry data.
}
