% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemR.R
\name{read_starr}
\alias{read_starr}
\title{Read .asc files produced by starr telemetry and put them in tidy tibbles}
\usage{
read_starr(
  raw_asc,
  meta_data_group,
  meta_data_xover = NULL,
  meta_data_sex = NULL,
  trim_na = F,
  trim_bad_probe = F
)
}
\arguments{
\item{raw_asc}{.asc file produced by starr telemetry system. Use "new format" and "date-time in left column only" when exporting .asc}

\item{meta_data_group}{data frame or tibble containing sample group information}

\item{meta_data_xover}{data frame or tibble containing sample corssover groups}

\item{meta_data_sex}{data frame or tibble containing sample sexes}

\item{trim_na}{logical: should function remove time points containing NA values?}

\item{trim_bad_probe}{logical: should function remove bad probes?}

\item{n_animals}{number of animals}

\item{n_measurements}{number of measurments taken}
}
\value{
a tidy telemetry tibble (tidy_telem)
}
\description{
Read .asc files produced by starr telemetry and put them in tidy tibbles
}
